% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_discrete_palette}
\alias{add_discrete_palette}
\title{Add discrete palette}
\usage{
add_discrete_palette(
  plot_plt,
  colours_chr = c("#de2d26", "#fc9272"),
  missing_1L_chr = "grey50",
  type_1L_chr = c("ggsci", "viridis"),
  what_1L_chr = "lancet"
)
}
\arguments{
\item{plot_plt}{Plot (a plot)}

\item{colours_chr}{Colours (a character vector), Default: c("#de2d26", "#fc9272")}

\item{missing_1L_chr}{Missing (a character vector of length one), Default: 'grey50'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("ggsci", "viridis")}

\item{what_1L_chr}{What (a character vector of length one), Default: 'lancet'}
}
\value{
Plot (a plot)
}
\description{
add_discrete_palette() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add discrete palette. The function returns Plot (a plot).
}
\keyword{internal}
