% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_participation_var}
\alias{add_participation_var}
\title{Add participation variable}
\usage{
add_participation_var(
  data_tb,
  id_var_nm_1L_chr = "fkClientID",
  fup_round_nmbr_1L_int = 2L
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}

\item{fup_round_nmbr_1L_int}{Follow-up round nmbr (an integer vector of length one), Default: 2}
}
\value{
Data (a tibble)
}
\description{
add_participation_var() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add participation variable. Function argument data_tb specifies the object to be updated. The function returns Data (a tibble).
}
