% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_wtd_dim_tots}
\alias{add_wtd_dim_tots}
\title{Add weighted dimension totals}
\usage{
add_wtd_dim_tots(
  unwtd_dim_tb,
  domain_items_ls,
  domain_unwtd_pfx_1L_chr,
  domain_wtd_pfx_1L_chr
)
}
\arguments{
\item{unwtd_dim_tb}{Unweighted dimension (a tibble)}

\item{domain_items_ls}{Domain items (a list)}

\item{domain_unwtd_pfx_1L_chr}{Domain unweighted prefix (a character vector of length one)}

\item{domain_wtd_pfx_1L_chr}{Domain weighted prefix (a character vector of length one)}
}
\value{
Weighted and unweighted dimension (a tibble)
}
\description{
add_wtd_dim_tots() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add weighted dimension totals. Function argument unwtd_dim_tb specifies the object to be updated. The function returns Weighted and unweighted dimension (a tibble).
}
\keyword{internal}
