% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_aqol_scrg_dict_r3.R
\docType{data}
\name{aqol_scrg_dict_r3}
\alias{aqol_scrg_dict_r3}
\title{Data dictionary for AQoL scoring}
\format{
An object of class \code{ready4_dictionary} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 59 rows and 4 columns.
}
\usage{
aqol_scrg_dict_r3
}
\description{
A data dictionary of the variables used in scoring AQoL 6D utility questionnaire responses.
}
\details{
A tibble

\describe{
\item{var_nm_chr}{Variable name (a character vector)}
\item{var_ctg_chr}{Variable category categories (a character vector)}
\item{var_desc_chr}{Variable description (a character vector)}
\item{var_type_chr}{Variable type (a character vector)}
}
}
\keyword{datasets}
