% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_assert.R
\name{assert_ds_is_valid}
\alias{assert_ds_is_valid}
\title{Assert dataset is valid}
\usage{
assert_ds_is_valid(
  data_tb,
  id_var_nm_1L_chr,
  round_var_nm_1L_chr,
  round_bl_val_1L_chr,
  msrmnt_date_var_nm_1L_chr = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one)}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one)}

\item{round_bl_val_1L_chr}{Round baseline value (a character vector of length one)}

\item{msrmnt_date_var_nm_1L_chr}{Measurement date variable name (a character vector of length one), Default: NULL}
}
\value{
No return value, called for side effects.
}
\description{
assert_ds_is_valid() is an Assert function that validates that an object conforms to required condition(s). If the object does not meet all required conditions, program execution will be stopped and an error message provided. Specifically, this function implements an algorithm to assert dataset is valid. The function is called for its side effects and does not return a value.
}
