% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_adult_aqol6dU}
\alias{calculate_adult_aqol6dU}
\title{Calculate adult Assessment of Quality of Life Six Dimension Health Utility}
\usage{
calculate_adult_aqol6dU(
  aqol6d_items_tb,
  prefix_1L_chr,
  coefs_lup_tb = NULL,
  dim_sclg_con_lup_tb = NULL,
  disvalues_lup_tb = NULL,
  itm_wrst_wts_lup_tb = NULL
)
}
\arguments{
\item{aqol6d_items_tb}{Assessment of Quality of Life Six Dimension items (a tibble)}

\item{prefix_1L_chr}{Prefix (a character vector of length one)}

\item{coefs_lup_tb}{Coefficients lookup table (a tibble), Default: NULL}

\item{dim_sclg_con_lup_tb}{Dimension scaling constant lookup table (a tibble), Default: NULL}

\item{disvalues_lup_tb}{Disvalues lookup table (a tibble), Default: NULL}

\item{itm_wrst_wts_lup_tb}{Item worst weights lookup table (a tibble), Default: NULL}
}
\value{
Assessment of Quality of Life Six Dimension Health Utility (a double vector)
}
\description{
calculate_adult_aqol6dU() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate adult assessment of quality of life six dimension health utility. The function returns Assessment of Quality of Life Six Dimension Health Utility (a double vector).
}
\keyword{internal}
