% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_characterize.R
\name{characterize-YouthvarsDescriptives}
\alias{characterize-YouthvarsDescriptives}
\alias{characterize,YouthvarsDescriptives-method}
\title{Characterize the features of a ready4 framework module (or sub-module)}
\usage{
\S4method{characterize}{YouthvarsDescriptives}(x, y_Ready4useDyad)
}
\arguments{
\item{x}{An object of class YouthvarsDescriptives}

\item{y_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
Descriptives (a data.frame)
}
\description{
characterize method applied to YouthvarsDescriptives
}
