% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_exhibit.R
\name{exhibit-YouthvarsSeries}
\alias{exhibit-YouthvarsSeries}
\alias{exhibit,YouthvarsSeries-method}
\alias{exhibit-YouthvarsProfile}
\alias{exhibit,YouthvarsProfile-method}
\title{Exhibit features of a dataset by printing them to the R console}
\usage{
\S4method{exhibit}{YouthvarsSeries}(
  x,
  captions_chr = NULL,
  mkdn_tbl_refs_chr = NULL,
  profile_idx_int = NA_integer_,
  output_type_1L_chr = "HTML",
  timepoints_int = c(1L, 2L),
  type_1L_chr = "characterize",
  what_1L_chr = "descriptives",
  ...
)

\S4method{exhibit}{YouthvarsProfile}(
  x,
  captions_chr = NULL,
  header_1L_chr = "",
  header_col_nms_chr = " ",
  mkdn_tbl_refs_chr = NULL,
  profile_idx_int = NA_integer_,
  output_type_1L_chr = "HTML",
  what_1L_chr = "descriptives",
  ...
)
}
\arguments{
\item{x}{An object of class YouthvarsProfile}

\item{captions_chr}{Captions (a character vector), Default: NULL}

\item{mkdn_tbl_refs_chr}{Markdown table references (a character vector), Default: NULL}

\item{profile_idx_int}{Profile index (an integer vector), Default: NA}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{timepoints_int}{Timepoints (an integer vector), Default: c(1L, 2L)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'characterize'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'descriptives'}

\item{...}{Additional arguments}

\item{header_1L_chr}{Header (a character vector of length one), Default: ''}

\item{header_col_nms_chr}{Header column names (a character vector), Default: ' '}
}
\description{
exhibit method applied to YouthvarsSeries

exhibit method applied to YouthvarsProfile
}
