% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_aqol6d_scrg_dss}
\alias{get_aqol6d_scrg_dss}
\title{Get Assessment of Quality of Life Six Dimension scoring datasets}
\usage{
get_aqol6d_scrg_dss()
}
\value{
Assessment of Quality of Life Six Dimension scoring datasets (a list)
}
\description{
get_aqol6d_scrg_dss() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get assessment of quality of life six dimension scoring datasets. The function returns Assessment of Quality of Life Six Dimension scoring datasets (a list).
}
\keyword{internal}
