% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_guide_box_lgd}
\alias{get_guide_box_lgd}
\title{Get guide box legend}
\usage{
get_guide_box_lgd(plot_plt)
}
\arguments{
\item{plot_plt}{Plot (a plot)}
}
\value{
Legend (a list)
}
\description{
get_guide_box_lgd() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get guide box legend. Function argument plot_plt specifies the where to look for the required object. The function returns Legend (a list).
}
\keyword{internal}
