% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_journal_palette_fn}
\alias{get_journal_palette_fn}
\title{Get journal palette function}
\usage{
get_journal_palette_fn(
  type_1L_chr = c("colour", "fill"),
  what_1L_chr = "lancet"
)
}
\arguments{
\item{type_1L_chr}{Type (a character vector of length one), Default: c("colour", "fill")}

\item{what_1L_chr}{What (a character vector of length one), Default: 'lancet'}
}
\value{
Journal palette (a function)
}
\description{
get_journal_palette_fn() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get journal palette function. The function returns Journal palette (a function).
}
\keyword{internal}
