% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_aqol6d_adol.R
\name{is_youthvars_aqol6d_adol}
\alias{is_youthvars_aqol6d_adol}
\title{Is youthvars Assessment of Quality of Life Six Dimension adolescent youthvars S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))}
\usage{
is_youthvars_aqol6d_adol(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the youthvars S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\description{
Check whether an object is a valid instance of the youthvars S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\details{
youthvars S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
