% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_bads.R
\name{is_youthvars_bads}
\alias{is_youthvars_bads}
\title{Is youthvars bads youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores}
\usage{
is_youthvars_bads(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\description{
Check whether an object is a valid instance of the youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\details{
youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
