% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_gad7.R
\name{is_youthvars_gad7}
\alias{is_youthvars_gad7}
\title{Is First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores}
\usage{
is_youthvars_gad7(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\description{
Check whether an object is a valid instance of the First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\details{
First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
