% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_k6.R
\name{is_youthvars_k6}
\alias{is_youthvars_k6}
\title{is youthvars K-6 youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores}
\usage{
is_youthvars_k6(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\description{
Check whether an object is a valid instance of the youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\details{
youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
