% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_sofas.R
\name{is_youthvars_sofas}
\alias{is_youthvars_sofas}
\title{Is youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)}
\usage{
is_youthvars_sofas(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\description{
Check whether an object is a valid instance of the youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\details{
youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
