% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cors_with_utl_tbl}
\alias{make_cors_with_utl_tbl}
\title{Make correlations with utility table}
\usage{
make_cors_with_utl_tbl(
  data_tb,
  ds_descvs_ls,
  dictionary_tb = NULL,
  cor_type_1L_chr = "pearson"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{ds_descvs_ls}{Dataset descriptives (a list)}

\item{dictionary_tb}{Dictionary (a tibble), Default: NULL}

\item{cor_type_1L_chr}{Correlation type (a character vector of length one), Default: 'pearson'}
}
\value{
Correlations with utility (a tibble)
}
\description{
make_cors_with_utl_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make correlations with utility table. The function returns Correlations with utility (a tibble).
}
\keyword{internal}
