% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_csnl_example_dict}
\alias{make_csnl_example_dict}
\title{Make cross-sectional example dictionary}
\usage{
make_csnl_example_dict(ds_tb)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}
}
\description{
make_csnl_example_dict() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cross-sectional example dictionary. The function is called for its side effects and does not return a value.
}
\keyword{internal}
