% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_descv_stats_tbl}
\alias{make_descv_stats_tbl}
\title{Make descriptive statistics table}
\usage{
make_descv_stats_tbl(
  data_tb,
  key_var_nm_1L_chr = "round",
  make_descv_stats_tbl,
  key_var_vals_chr,
  variable_nms_chr,
  test_1L_lgl = F
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{key_var_nm_1L_chr}{Key variable name (a character vector of length one), Default: 'round'}

\item{make_descv_stats_tbl}{PARAM_DESCRIPTION}

\item{key_var_vals_chr}{Key variable values (a character vector)}

\item{variable_nms_chr}{Variable names (a character vector)}

\item{test_1L_lgl}{Test (a logical vector of length one), Default: F}
}
\value{
Descriptive statistics table (a tibble)
}
\description{
make_descv_stats_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make descriptive statistics table. The function returns Descriptive statistics table (a tibble).
}
\keyword{internal}
