% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_final_rpln_ds_dict}
\alias{make_final_rpln_ds_dict}
\title{Make final rpln dataset dictionary}
\usage{
make_final_rpln_ds_dict(
  seed_dictionary_tb = NULL,
  additions_tb = NULL,
  utl_unwtd_var_nm_1L_chr = "aqol6d_total_c"
)
}
\arguments{
\item{seed_dictionary_tb}{Seed dictionary (a tibble), Default: NULL}

\item{additions_tb}{Additions (a tibble), Default: NULL}

\item{utl_unwtd_var_nm_1L_chr}{Utility unwtd variable name (a character vector of length one), Default: 'aqol6d_total_c'}
}
\value{
Dictionary (a tibble)
}
\description{
make_final_rpln_ds_dict() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make final rpln dataset dictionary. The function returns Dictionary (a tibble).
}
