% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_formula}
\alias{make_formula}
\title{Make formula}
\usage{
make_formula(
  depnt_var_nm_1L_chr,
  predictors_chr,
  environment_env = parent.frame()
)
}
\arguments{
\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one)}

\item{predictors_chr}{Predictors (a character vector)}

\item{environment_env}{Environment (an environment), Default: parent.frame()}
}
\value{
formula_fml (An object)
}
\description{
make_formula() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make formula. The function is called for its side effects and does not return a value.
}
