% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_chu9d_adolaus.R
\name{make_new_youthvars_chu9d_adolaus}
\alias{make_new_youthvars_chu9d_adolaus}
\title{make new youthvars chu9d adolaus youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)}
\usage{
make_new_youthvars_chu9d_adolaus(x)
}
\arguments{
\item{x}{A prototype for the youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)}
}
\value{
An unvalidated instance of the youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)
}
\description{
Create a new unvalidated instance of the youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)
}
\details{
youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension - Australian Adolescent Scoring (CHU-9D Australian Adolescent)
}
\keyword{internal}
