% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_k10_aus.R
\name{make_new_youthvars_k10_aus}
\alias{make_new_youthvars_k10_aus}
\title{make new youthvars K10 Australia youthvars ready4 sub-module (S3 class) for Kessler Psychological Distress Scale (K10) - Australian Scoring System scores}
\usage{
make_new_youthvars_k10_aus(x)
}
\arguments{
\item{x}{A prototype for the youthvars ready4 sub-module (S3 class) for Kessler Psychological Distress Scale (K10) - Australian Scoring System scores}
}
\value{
An unvalidated instance of the youthvars ready4 sub-module (S3 class) for Kessler Psychological Distress Scale (K10) - Australian Scoring System scores
}
\description{
Create a new unvalidated instance of the youthvars ready4 sub-module (S3 class) for Kessler Psychological Distress Scale (K10) - Australian Scoring System scores
}
\details{
youthvars ready4 sub-module (S3 class) for Kessler Psychological Distress Scale (K10) - Australian Scoring System scores
}
\keyword{internal}
