% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_oasis.R
\name{make_new_youthvars_oasis}
\alias{make_new_youthvars_oasis}
\title{Make new First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores}
\usage{
make_new_youthvars_oasis(x)
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores}
}
\value{
An unvalidated instance of the First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\description{
Create a new unvalidated instance of the First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\details{
First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
