% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_phq9.R
\name{make_new_youthvars_phq9}
\alias{make_new_youthvars_phq9}
\title{Make new youthvars S3 class for Patient Health Questionnaire (PHQ-9) scores}
\usage{
make_new_youthvars_phq9(x)
}
\arguments{
\item{x}{A prototype for the youthvars S3 class for Patient Health Questionnaire (PHQ-9) scores}
}
\value{
An unvalidated instance of the youthvars S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\description{
Create a new unvalidated instance of the youthvars S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\details{
youthvars S3 class for Patient Health Questionnaire (PHQ-9) scores
}
