% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_paths_ls}
\alias{make_paths_ls}
\title{Make paths}
\usage{
make_paths_ls(params_ls)
}
\arguments{
\item{params_ls}{Params (a list)}
}
\value{
Paths (a list)
}
\description{
make_paths_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make paths list. The function returns Paths (a list).
}
\keyword{internal}
