% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_predr_pars_and_cors_tbl}
\alias{make_predr_pars_and_cors_tbl}
\title{Make predictor pars and correlations table}
\usage{
make_predr_pars_and_cors_tbl(
  data_tb,
  ds_descvs_ls,
  descv_tbl_ls,
  dictionary_tb,
  nbr_of_digits_1L_int = 2L,
  predictors_lup = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{ds_descvs_ls}{Dataset descriptives (a list)}

\item{descv_tbl_ls}{Descriptive table (a list)}

\item{dictionary_tb}{Dictionary (a tibble)}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}

\item{predictors_lup}{Predictors (a lookup table), Default: NULL}
}
\value{
Predictor pars and correlations (a tibble)
}
\description{
make_predr_pars_and_cors_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make predictor pars and correlations table. The function returns Predictor pars and correlations (a tibble).
}
\keyword{internal}
