% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_aqol6d_adol.R
\name{make_pt_youthvars_aqol6d_adol}
\alias{make_pt_youthvars_aqol6d_adol}
\title{make prototype youthvars Assessment of Quality of Life Six Dimension adolescent youthvars S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))}
\usage{
make_pt_youthvars_aqol6d_adol()
}
\value{
A prototype for youthvars S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\description{
Create a new prototype for the youthvars S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\details{
youthvars S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
