% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_bads.R
\name{make_pt_youthvars_bads}
\alias{make_pt_youthvars_bads}
\title{Make prototype youthvars bads youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores}
\usage{
make_pt_youthvars_bads()
}
\value{
A prototype for youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\description{
Create a new prototype for the youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\details{
youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
