% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_k6.R
\name{make_pt_youthvars_k6}
\alias{make_pt_youthvars_k6}
\title{Make prototype youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores}
\usage{
make_pt_youthvars_k6()
}
\value{
A prototype for youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\description{
Create a new prototype for the youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\details{
youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
