% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_oasis.R
\name{make_pt_youthvars_oasis}
\alias{make_pt_youthvars_oasis}
\title{Make prototype First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores}
\usage{
make_pt_youthvars_oasis()
}
\value{
A prototype for First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\description{
Create a new prototype for the First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\details{
First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
