% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_phq9.R
\name{make_pt_youthvars_phq9}
\alias{make_pt_youthvars_phq9}
\title{Make prototype youthvars PHQ-9 youthvars S3 class for Patient Health Questionnaire (PHQ-9) scores}
\usage{
make_pt_youthvars_phq9()
}
\value{
A prototype for youthvars S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\description{
Create a new prototype for the youthvars S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\details{
youthvars S3 class for Patient Health Questionnaire (PHQ-9) scores
}
