% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_scared.R
\name{make_pt_youthvars_scared}
\alias{make_pt_youthvars_scared}
\title{Make prototype youthvars scared youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores}
\usage{
make_pt_youthvars_scared()
}
\value{
A prototype for youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\description{
Create a new prototype for the youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\details{
youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
