% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_sofas.R
\name{make_pt_youthvars_sofas}
\alias{make_pt_youthvars_sofas}
\title{make prototype youthvars sofas youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)}
\usage{
make_pt_youthvars_sofas()
}
\value{
A prototype for youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\description{
Create a new prototype for the youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\details{
youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
