% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sub_ttl_plts}
\alias{make_sub_ttl_plts}
\title{Make sub ttl plots}
\usage{
make_sub_ttl_plts(
  data_tb,
  col_nms_chr,
  plot_rows_cols_pair_int,
  round_var_nm_1L_chr = "round",
  make_log_log_tfmn_1L_lgl = F,
  heights_int
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{col_nms_chr}{Column names (a character vector)}

\item{plot_rows_cols_pair_int}{Plot rows columns pair (an integer vector)}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{make_log_log_tfmn_1L_lgl}{Make log log transformation (a logical vector of length one), Default: F}

\item{heights_int}{Heights (an integer vector)}
}
\value{
Composite (a plot)
}
\description{
make_sub_ttl_plts() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sub ttl plots. The function returns Composite (a plot).
}
