% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_tableby_ls}
\alias{make_tableby_ls}
\title{Make tableby}
\usage{
make_tableby_ls(data_tb, key_var_nm_1L_chr, variable_nms_chr, test_1L_lgl = F)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{key_var_nm_1L_chr}{Key variable name (a character vector of length one)}

\item{variable_nms_chr}{Variable names (a character vector)}

\item{test_1L_lgl}{Test (a logical vector of length one), Default: F}
}
\value{
Tableby (a list)
}
\description{
make_tableby_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make tableby list. The function returns Tableby (a list).
}
\keyword{internal}
