% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_var_by_round_plt}
\alias{make_var_by_round_plt}
\title{Make variable by round}
\usage{
make_var_by_round_plt(
  data_tb,
  var_nm_1L_chr,
  round_var_nm_1L_chr = "round",
  x_label_1L_chr,
  y_label_1L_chr = "Percentage",
  y_scale_scl_fn = scales::percent,
  label_fill_1L_chr = "Data collection"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{var_nm_1L_chr}{Variable name (a character vector of length one)}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{x_label_1L_chr}{X label (a character vector of length one)}

\item{y_label_1L_chr}{Y label (a character vector of length one), Default: 'Percentage'}

\item{y_scale_scl_fn}{Y scale scale (a function), Default: scales::percent}

\item{label_fill_1L_chr}{Label fill (a character vector of length one), Default: 'Data collection'}
}
\value{
Variable by round (a plot)
}
\description{
make_var_by_round_plt() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make variable by round plot. The function returns Variable by round (a plot).
}
