% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-YouthvarsProfile}
\alias{manufacture-YouthvarsProfile}
\alias{manufacture,YouthvarsProfile-method}
\alias{manufacture-YouthvarsSeries}
\alias{manufacture,YouthvarsSeries-method}
\title{Manufacture a new object}
\usage{
\S4method{manufacture}{YouthvarsProfile}(
  x,
  element_nm_1L_chr = "overall",
  nbr_of_digits_1L_int = 3L,
  profile_chr = character(0),
  what_1L_chr = "descriptives_ls",
  ...
)

\S4method{manufacture}{YouthvarsSeries}(
  x,
  compare_by_time_chr = NA_character_,
  compare_by_time_with_test_chr = NA_character_,
  compare_ptcpn_chr = NA_character_,
  compare_ptcpn_with_test_chr = NA_character_,
  nbr_of_digits_1L_int = 3L,
  timepoints_int = c(1L, 2L),
  what_1L_chr = "descriptives_ls",
  ...
)
}
\arguments{
\item{x}{An object of class YouthvarsSeries}

\item{element_nm_1L_chr}{Element name (a character vector of length one), Default: 'overall'}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 3}

\item{profile_chr}{Profile (a character vector), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'descriptives_ls'}

\item{...}{Additional arguments}

\item{compare_by_time_chr}{Compare by time (a character vector), Default: 'NA'}

\item{compare_by_time_with_test_chr}{Compare by time with test (a character vector), Default: 'NA'}

\item{compare_ptcpn_chr}{Compare ptcpn (a character vector), Default: 'NA'}

\item{compare_ptcpn_with_test_chr}{Compare ptcpn with test (a character vector), Default: 'NA'}

\item{timepoints_int}{Timepoints (an integer vector), Default: c(1L, 2L)}
}
\value{
Object (an output object of multiple potential types)

Object (an output object of multiple potential types)
}
\description{
manufacture method applied to YouthvarsProfile

manufacture method applied to YouthvarsSeries
}
