% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_descriptives}
\alias{plot_descriptives}
\title{Plot descriptives}
\usage{
plot_descriptives(
  X_YouthvarsProfile,
  what_1L_chr,
  legend_1L_lgl = FALSE,
  missing_1L_chr = "grey50",
  palette_1L_chr = "lancet",
  type_1L_chr = c("ggsci", "viridis"),
  x_ready4show_correspondences = ready4show::ready4show_correspondences(),
  y_label_1L_chr = "Count"
)
}
\arguments{
\item{X_YouthvarsProfile}{PARAM_DESCRIPTION}

\item{what_1L_chr}{What (a character vector of length one)}

\item{legend_1L_lgl}{Legend (a logical vector of length one), Default: FALSE}

\item{missing_1L_chr}{Missing (a character vector of length one), Default: 'grey50'}

\item{palette_1L_chr}{Palette (a character vector of length one), Default: 'lancet'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("ggsci", "viridis")}

\item{x_ready4show_correspondences}{PARAM_DESCRIPTION, Default: ready4show::ready4show_correspondences()}

\item{y_label_1L_chr}{Y label (a character vector of length one), Default: 'Count'}
}
\value{
Plot (a plot)
}
\description{
plot_descriptives() is a Plot function that plots data. Specifically, this function implements an algorithm to plot descriptives. The function returns Plot (a plot).
}
