% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_ratify.R
\name{ratify-YouthvarsProfile}
\alias{ratify-YouthvarsProfile}
\alias{ratify,YouthvarsProfile-method}
\alias{ratify-YouthvarsSeries}
\alias{ratify,YouthvarsSeries-method}
\title{Ratify that a dataset meets validity criteria}
\usage{
\S4method{ratify}{YouthvarsProfile}(x, ...)

\S4method{ratify}{YouthvarsSeries}(x, timepoints_int = c(1L, 2L), type_1L_chr = "two_timepoints", ...)
}
\arguments{
\item{x}{An object of class YouthvarsSeries}

\item{...}{Additional arguments}

\item{timepoints_int}{Timepoints (an integer vector), Default: c(1L, 2L)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'two_timepoints'}
}
\value{
x (An object of class YouthvarsProfile)

x (An object of class YouthvarsSeries)
}
\description{
ratify method applied to YouthvarsProfile

ratify method applied to YouthvarsSeries
}
