% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew-YouthvarsDescriptives}
\alias{renew-YouthvarsDescriptives}
\alias{renew,YouthvarsDescriptives-method}
\alias{renew-YouthvarsProfile}
\alias{renew,YouthvarsProfile-method}
\title{Renew (update) an object}
\usage{
\S4method{renew}{YouthvarsDescriptives}(x, type_1L_chr = "characterize", ...)

\S4method{renew}{YouthvarsProfile}(x, type_1L_chr = "characterize")
}
\arguments{
\item{x}{An object of class YouthvarsProfile}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'characterize'}

\item{...}{Additional arguments}
}
\value{
x (An object of class YouthvarsDescriptives)

x (An object of class YouthvarsProfile)
}
\description{
renew method applied to YouthvarsDescriptives

renew method applied to YouthvarsProfile
}
