% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_csnl_example_ds}
\alias{transform_csnl_example_ds}
\title{Transform cross-sectional example dataset}
\usage{
transform_csnl_example_ds(ds_df)
}
\arguments{
\item{ds_df}{Dataset (a data.frame)}
}
\value{
Dataset (a tibble)
}
\description{
transform_csnl_example_ds() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform cross-sectional example dataset. The function returns Dataset (a tibble).
}
\keyword{internal}
