% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_ds_for_item_plt}
\alias{transform_ds_for_item_plt}
\title{Transform dataset for item plot}
\usage{
transform_ds_for_item_plt(
  data_tb,
  var_nm_1L_chr,
  round_var_nm_1L_chr = "round"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{var_nm_1L_chr}{Variable name (a character vector of length one)}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}
}
\value{
Transformed data (a tibble)
}
\description{
transform_ds_for_item_plt() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dataset for item plot. The function returns Transformed data (a tibble).
}
\keyword{internal}
