% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_raw_ds_for_analysis}
\alias{transform_raw_ds_for_analysis}
\title{Transform raw dataset for analysis}
\usage{
transform_raw_ds_for_analysis(raw_ds_tb)
}
\arguments{
\item{raw_ds_tb}{Raw dataset (a tibble)}
}
\value{
Transformed dataset (a tibble)
}
\description{
transform_raw_ds_for_analysis() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform raw dataset for analysis. Function argument raw_ds_tb specifies the object to be updated. The function returns Transformed dataset (a tibble).
}
