% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_aqol6d_adol.R
\name{validate_youthvars_aqol6d_adol}
\alias{validate_youthvars_aqol6d_adol}
\title{Validate First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))}
\usage{
validate_youthvars_aqol6d_adol(x)
}
\arguments{
\item{x}{An unvalidated instance of the First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))}
}
\value{
A prototpe for First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\description{
Validate an instance of the First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\details{
First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
