% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_chu9d_adolaus.R
\name{validate_youthvars_chu9d_adolaus}
\alias{validate_youthvars_chu9d_adolaus}
\title{validate youthvars chu9d adolaus youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)}
\usage{
validate_youthvars_chu9d_adolaus(x)
}
\arguments{
\item{x}{An unvalidated instance of the youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)}
}
\value{
A prototpe for youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)
}
\description{
Validate an instance of the youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)
}
\details{
youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension - Australian Adolescent Scoring (CHU-9D Australian Adolescent)
}
\keyword{internal}
