% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_oasis.R
\name{validate_youthvars_oasis}
\alias{validate_youthvars_oasis}
\title{Validate First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores}
\usage{
validate_youthvars_oasis(x)
}
\arguments{
\item{x}{An unvalidated instance of the First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores}
}
\value{
A prototpe for First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\description{
Validate an instance of the First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\details{
First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
