% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_all_outp_dirs}
\alias{write_all_outp_dirs}
\title{Write all output directories}
\usage{
write_all_outp_dirs(paths_ls)
}
\arguments{
\item{paths_ls}{Paths (a list)}
}
\value{
Paths (a list)
}
\description{
write_all_outp_dirs() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write all output directories. The function returns Paths (a list).
}
\keyword{internal}
