% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_descv_tbls}
\alias{write_descv_tbls}
\title{Write descriptive tables}
\usage{
write_descv_tbls(
  data_tb,
  ds_descvs_ls,
  predictors_lup,
  descv_outp_dir_1L_chr,
  nbr_of_digits_1L_int = 2
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{ds_descvs_ls}{Dataset descriptives (a list)}

\item{predictors_lup}{Predictors (a lookup table)}

\item{descv_outp_dir_1L_chr}{Descriptive output directory (a character vector of length one)}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}
}
\value{
Descriptive table (a list)
}
\description{
write_descv_tbls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write descriptive tables. The function returns Descriptive table (a list).
}
\keyword{internal}
