% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_qstn_resp_plts}
\alias{write_qstn_resp_plts}
\title{Write qstn resp plots}
\usage{
write_qstn_resp_plts(
  data_tb,
  col_nms_chr,
  lbl_nms_chr,
  plot_rows_cols_pair_int,
  heights_int,
  round_var_nm_1L_chr = "round"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{col_nms_chr}{Column names (a character vector)}

\item{lbl_nms_chr}{Label names (a character vector)}

\item{plot_rows_cols_pair_int}{Plot rows columns pair (an integer vector)}

\item{heights_int}{Heights (an integer vector)}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}
}
\value{
Composite (a plot)
}
\description{
write_qstn_resp_plts() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write qstn resp plots. The function returns Composite (a plot).
}
\keyword{internal}
