% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_bads.R
\name{youthvars_bads}
\alias{youthvars_bads}
\alias{make_pt_youthvars_bads}
\alias{is_youthvars_bads}
\title{youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores}
\usage{
youthvars_bads(x = make_pt_youthvars_bads())

make_pt_youthvars_bads()

is_youthvars_bads(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A validated instance of the youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores

A prototype for youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores

A logical value, TRUE if a valid instance of the youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\description{
Create a new valid instance of the youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\details{
youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
