% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_chu9d_adolaus.R
\name{youthvars_chu9d_adolaus}
\alias{youthvars_chu9d_adolaus}
\alias{make_pt_youthvars_chu9d_adolaus}
\alias{is_youthvars_chu9d_adolaus}
\title{youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)}
\usage{
youthvars_chu9d_adolaus(x = make_pt_youthvars_chu9d_adolaus())

make_pt_youthvars_chu9d_adolaus()

is_youthvars_chu9d_adolaus(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A validated instance of the youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)

A prototype for youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)

A logical value, TRUE if a valid instance of the youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)
}
\description{
Create a new valid instance of the youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)
}
\details{
youthvars ready4 sub-module (S3 class) for Child Health Utility Nine Dimension Health Utility - Australian Adolescent Scoring (CHU-9D Australian Adolescent)
}
