% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_k10_aus.R
\name{youthvars_k10_aus}
\alias{youthvars_k10_aus}
\alias{make_pt_youthvars_k10_aus}
\alias{is_youthvars_k10_aus}
\title{youthvars ready4 sub-module (S3 class) for Kessler Psychological Distress Scale (K10) - Australian Scoring System scores}
\usage{
youthvars_k10_aus(x = make_pt_youthvars_k10_aus())

make_pt_youthvars_k10_aus()

is_youthvars_k10_aus(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A validated instance of the youthvars ready4 sub-module (S3 class) for Kessler Psychological Distress Scale (K10) - Australian Scoring System scores

A prototype for youthvars ready4 sub-module (S3 class) for Kessler Psychological Distress Scale (K10) - Australian Scoring System scores

A logical value, TRUE if a valid instance of the youthvars ready4 sub-module (S3 class) for Kessler Psychological Distress Scale (K10) - Australian Scoring System scores
}
\description{
Create a new valid instance of the youthvars ready4 sub-module (S3 class) for Kessler Psychological Distress Scale (K10) - Australian Scoring System scores
}
\details{
youthvars ready4 sub-module (S3 class) for Kessler Psychological Distress Scale (K10) - Australian Scoring System scores
}
