% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_scared.R
\name{youthvars_scared}
\alias{youthvars_scared}
\alias{make_pt_youthvars_scared}
\alias{is_youthvars_scared}
\title{youthvars ready4 sub-module (S3 class) for Screen for Child Anxiety Related Disorders (SCARED) scores}
\usage{
youthvars_scared(x = make_pt_youthvars_scared())

make_pt_youthvars_scared()

is_youthvars_scared(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A validated instance of the youthvars ready4 sub-module (S3 class) for Screen for Child Anxiety Related Disorders (SCARED) scores

A prototype for youthvars ready4 sub-module (S3 class) for Screen for Child Anxiety Related Disorders (SCARED) scores

A logical value, TRUE if a valid instance of the youthvars ready4 sub-module (S3 class) for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\description{
Create a new valid instance of the youthvars ready4 sub-module (S3 class) for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\details{
youthvars ready4 sub-module (S3 class) for Screen for Child Anxiety Related Disorders (SCARED) scores
}
