% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_sofas.R
\name{youthvars_sofas}
\alias{youthvars_sofas}
\title{First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)}
\usage{
youthvars_sofas(x = make_pt_youthvars_sofas())
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS), Default: make_pt_youthvars_sofas()}
}
\value{
A validated instance of the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\description{
Create a new valid instance of the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\details{
First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
