% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_YouthvarsDescriptives.R
\docType{class}
\name{YouthvarsDescriptives-class}
\alias{YouthvarsDescriptives-class}
\alias{YouthvarsDescriptives}
\title{YouthvarsDescriptives}
\description{
Metadata about descriptive statistics to be generated.
}
\section{Slots}{

\describe{
\item{\code{descriptives_df}}{Descriptives (a data.frame)}

\item{\code{ds_tfmn_ls}}{Dataset transformation (a list)}

\item{\code{key_var_nm_1L_chr}}{Key variable name (a character vector of length one)}

\item{\code{key_var_vals_chr}}{Key variable values (a character vector)}

\item{\code{nbr_of_digits_1L_int}}{Number of digits (an integer vector of length one)}

\item{\code{profiled_vars_chr}}{Profiled variables (a character vector)}

\item{\code{sections_as_row_1L_lgl}}{Sections as row (a logical vector of length one)}

\item{\code{test_1L_lgl}}{Test (a logical vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

