% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_YouthvarsProfile.R
\docType{class}
\name{YouthvarsProfile-class}
\alias{YouthvarsProfile-class}
\alias{YouthvarsProfile}
\title{YouthvarsProfile}
\description{
A dataset and its associated dictionary, descriptive statistics and metadata.
}
\section{Slots}{

\describe{
\item{\code{a_Ready4useDyad}}{NO MATCH}

\item{\code{descriptives_ls}}{Descriptives (a list)}

\item{\code{id_var_nm_1L_chr}}{Identity variable name (a character vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

