% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_YouthvarsSeries.R
\docType{class}
\name{YouthvarsSeries-class}
\alias{YouthvarsSeries-class}
\alias{YouthvarsSeries}
\title{YouthvarsSeries}
\description{
A longitudinal dataset and its associated dictionary, descriptive statistics and metadata.
}
\section{Slots}{

\describe{
\item{\code{participation_var_1L_chr}}{Participation variable (a character vector of length one)}

\item{\code{timepoint_vals_chr}}{Timepoint values (a character vector)}

\item{\code{timepoint_var_nm_1L_chr}}{Timepoint variable name (a character vector of length one)}

\item{\code{a_Ready4useDyad}}{(an instance of the Ready4useDyad class)}

\item{\code{descriptives_ls}}{Descriptives (a list)}

\item{\code{id_var_nm_1L_chr}}{Identity variable name (a character vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

