% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_adol6d_scores}
\alias{add_adol6d_scores}
\title{Add Assessment of Quality of Life Six Dimension (Adolescent version) scores}
\usage{
add_adol6d_scores(
  unscored_aqol_tb,
  aqol6d_scrg_dss_ls = NULL,
  id_var_nm_1L_chr = "fkClientID",
  prefix_1L_chr = "aqol6d_q",
  total_aqol_var_nm_1L_chr = "aqol6d_total_c",
  wtd_aqol_var_nm_1L_chr = "aqol6d_total_w"
)
}
\arguments{
\item{unscored_aqol_tb}{Unscored Assessment of Quality of Life (a tibble)}

\item{aqol6d_scrg_dss_ls}{Assessment of Quality of Life Six Dimension scoring datasets (a list), Default: NULL}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'aqol6d_q'}

\item{total_aqol_var_nm_1L_chr}{Total Assessment of Quality of Life variable name (a character vector of length one), Default: 'aqol6d_total_c'}

\item{wtd_aqol_var_nm_1L_chr}{Weighted Assessment of Quality of Life variable name (a character vector of length one), Default: 'aqol6d_total_w'}
}
\value{
Transformed Assessment of Quality of Life (a tibble)
}
\description{
add_adol6d_scores() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add assessment of quality of life six dimension (adolescent version) scores. The function returns Transformed Assessment of Quality of Life (a tibble).
}
\keyword{internal}
