% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_adult_aqol6dU}
\alias{calculate_adult_aqol6dU}
\title{Calculate adult Assessment of Quality of Life Six Dimension Health Utility}
\usage{
calculate_adult_aqol6dU(
  aqol6d_items_tb,
  prefix_1L_chr,
  aqol6d_scrg_dss_ls = NULL
)
}
\arguments{
\item{aqol6d_items_tb}{Assessment of Quality of Life Six Dimension items (a tibble)}

\item{prefix_1L_chr}{Prefix (a character vector of length one)}

\item{aqol6d_scrg_dss_ls}{Assessment of Quality of Life Six Dimension scoring datasets (a list), Default: NULL}
}
\value{
Assessment of Quality of Life Six Dimension Health Utility (a double vector)
}
\description{
calculate_adult_aqol6dU() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate adult assessment of quality of life six dimension health utility. The function returns Assessment of Quality of Life Six Dimension Health Utility (a double vector).
}
\keyword{internal}
