% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_depict.R
\name{depict-YouthvarsSeries}
\alias{depict-YouthvarsSeries}
\alias{depict,YouthvarsSeries-method}
\alias{depict-YouthvarsProfile}
\alias{depict,YouthvarsProfile-method}
\title{Depict (plot) features of model module data}
\usage{
\S4method{depict}{YouthvarsSeries}(
  x,
  var_nms_chr,
  type_1L_chr = "by_time",
  labels_chr = NA_character_,
  label_fill_1L_chr = "Data collection",
  y_label_1L_chr = "Percentage",
  y_scale_scl_fn = scales::percent,
  ...
)

\S4method{depict}{YouthvarsProfile}(
  x,
  var_nms_chr,
  labels_chr = NA_character_,
  y_label_1L_chr = "Percentage",
  y_scale_scl_fn = scales::percent,
  ...
)
}
\arguments{
\item{x}{An object of class YouthvarsProfile}

\item{var_nms_chr}{Variable names (a character vector)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'by_time'}

\item{labels_chr}{Labels (a character vector), Default: 'NA'}

\item{label_fill_1L_chr}{Label fill (a character vector of length one), Default: 'Data collection'}

\item{y_label_1L_chr}{Y label (a character vector of length one), Default: 'Percentage'}

\item{y_scale_scl_fn}{Y scale scale (a function), Default: scales::percent}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects.

No return value, called for side effects.
}
\description{
depict method applied to YouthvarsSeries

depict method applied to YouthvarsProfile
}
