% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_oasis.R
\name{is_youthvars_oasis}
\alias{is_youthvars_oasis}
\title{Is youthvars oasis youthvars S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores}
\usage{
is_youthvars_oasis(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the youthvars S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\description{
Check whether an object is a valid instance of the youthvars S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\details{
youthvars S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
