% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_item_plt}
\alias{make_item_plt}
\title{Make item plot}
\usage{
make_item_plt(
  tfd_data_tb,
  var_nm_1L_chr,
  x_label_1L_chr,
  fill_label_1L_chr = "Data collection",
  legend_position_1L_chr = "none",
  round_var_nm_1L_chr = "round",
  sngl_round_lbl_1L_chr = "n",
  sngl_round_var_nm_1L_chr = "n",
  use_bw_theme_1L_lgl = F,
  y_label_1L_chr = "Percentage",
  y_scale_scl_fn = NULL
)
}
\arguments{
\item{tfd_data_tb}{Transformed data (a tibble)}

\item{var_nm_1L_chr}{Variable name (a character vector of length one)}

\item{x_label_1L_chr}{X label (a character vector of length one)}

\item{fill_label_1L_chr}{Fill label (a character vector of length one), Default: 'Data collection'}

\item{legend_position_1L_chr}{Legend position (a character vector of length one), Default: 'none'}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{sngl_round_lbl_1L_chr}{Single round label (a character vector of length one), Default: 'n'}

\item{sngl_round_var_nm_1L_chr}{Single round variable name (a character vector of length one), Default: 'n'}

\item{use_bw_theme_1L_lgl}{Use black and white theme (a logical vector of length one), Default: F}

\item{y_label_1L_chr}{Y label (a character vector of length one), Default: 'Percentage'}

\item{y_scale_scl_fn}{Y scale scale (a function), Default: NULL}
}
\value{
Item (a plot)
}
\description{
make_item_plt() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make item plot. The function returns Item (a plot).
}
\keyword{internal}
