% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_scared.R
\name{make_new_youthvars_scared}
\alias{make_new_youthvars_scared}
\title{Make new youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores}
\usage{
make_new_youthvars_scared(x)
}
\arguments{
\item{x}{A prototype for the youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores}
}
\value{
An unvalidated instance of the youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\description{
Create a new unvalidated instance of the youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\details{
youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
