% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_sofas.R
\name{make_new_youthvars_sofas}
\alias{make_new_youthvars_sofas}
\title{Make new First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)}
\usage{
make_new_youthvars_sofas(x)
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)}
}
\value{
An unvalidated instance of the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\description{
Create a new unvalidated instance of the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\details{
First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
