% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_gad7.R
\name{make_pt_youthvars_gad7}
\alias{make_pt_youthvars_gad7}
\title{Make prototype youthvars S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores}
\usage{
make_pt_youthvars_gad7()
}
\value{
A prototype for youthvars S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\description{
Create a new prototype for the youthvars S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\details{
youthvars S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
