% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_subtotal_plt}
\alias{make_subtotal_plt}
\title{Make subtotal}
\usage{
make_subtotal_plt(
  data_tb,
  var_nm_1L_chr,
  round_var_nm_1L_chr = "round",
  x_label_1L_chr,
  y_label_1L_chr = "Percentage",
  y_scale_scl_fn = scales::percent,
  use_bw_theme_1L_lgl = T,
  legend_position_1L_chr = "none",
  label_fill_1L_chr = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{var_nm_1L_chr}{Variable name (a character vector of length one)}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{x_label_1L_chr}{X label (a character vector of length one)}

\item{y_label_1L_chr}{Y label (a character vector of length one), Default: 'Percentage'}

\item{y_scale_scl_fn}{Y scale scale (a function), Default: scales::percent}

\item{use_bw_theme_1L_lgl}{Use black and white theme (a logical vector of length one), Default: T}

\item{legend_position_1L_chr}{Legend position (a character vector of length one), Default: 'none'}

\item{label_fill_1L_chr}{Label fill (a character vector of length one), Default: NULL}
}
\value{
Subtotal (a plot)
}
\description{
make_subtotal_plt() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make subtotal plot. The function returns Subtotal (a plot).
}
\keyword{internal}
