% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_tableby_cntrls}
\alias{make_tableby_cntrls}
\title{Make tableby cntrls}
\usage{
make_tableby_cntrls(test_1L_lgl = F)
}
\arguments{
\item{test_1L_lgl}{Test (a logical vector of length one), Default: F}
}
\value{
Tableby cntrls (a list)
}
\description{
make_tableby_cntrls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make tableby cntrls. The function returns Tableby cntrls (a list).
}
\keyword{internal}
