% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renewSlot.R
\name{renewSlot-YouthvarsSeries}
\alias{renewSlot-YouthvarsSeries}
\alias{renewSlot,YouthvarsSeries-method}
\title{renewSlot}
\usage{
\S4method{renewSlot}{YouthvarsSeries}(
  x,
  slot_nm_1L_chr = "descriptives_ls",
  nbr_of_digits_1L_int = 3L,
  profiled_vars_ls = NULL,
  timepoints_int = c(1L, 2L)
)
}
\arguments{
\item{x}{An object of class YouthvarsSeries}

\item{slot_nm_1L_chr}{Slot name (a character vector of length one), Default: 'descriptives_ls'}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 3}

\item{profiled_vars_ls}{Profiled variables (a list), Default: NULL}

\item{timepoints_int}{Timepoints (an integer vector), Default: c(1L, 2L)}
}
\value{
x (An object of class YouthvarsSeries)
}
\description{
renewSlot method applied to YouthvarsSeries
}
