% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_bads.R
\name{validate_youthvars_bads}
\alias{validate_youthvars_bads}
\title{Validate youthvars bads youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores}
\usage{
validate_youthvars_bads(x)
}
\arguments{
\item{x}{An unvalidated instance of the youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores}
}
\value{
A prototpe for youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\description{
Validate an instance of the youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\details{
youthvars S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
