% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_gad7.R
\name{validate_youthvars_gad7}
\alias{validate_youthvars_gad7}
\title{validate youthvars GAD-7 youthvars ready4 sub-module (S3 class) for Generalised Anxiety Disorder Scale (GAD-7) scores}
\usage{
validate_youthvars_gad7(x)
}
\arguments{
\item{x}{An unvalidated instance of the youthvars ready4 sub-module (S3 class) for Generalised Anxiety Disorder Scale (GAD-7) scores}
}
\value{
A prototpe for youthvars ready4 sub-module (S3 class) for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\description{
Validate an instance of the youthvars ready4 sub-module (S3 class) for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\details{
youthvars ready4 sub-module (S3 class) for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\keyword{internal}
