% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_sofas.R
\name{validate_youthvars_sofas}
\alias{validate_youthvars_sofas}
\title{validate youthvars sofas youthvars ready4 sub-module (S3 class) for Social and Occupational Functioning Assessment Scale (SOFAS)}
\usage{
validate_youthvars_sofas(x)
}
\arguments{
\item{x}{An unvalidated instance of the youthvars ready4 sub-module (S3 class) for Social and Occupational Functioning Assessment Scale (SOFAS)}
}
\value{
A prototpe for youthvars ready4 sub-module (S3 class) for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\description{
Validate an instance of the youthvars ready4 sub-module (S3 class) for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\details{
youthvars ready4 sub-module (S3 class) for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\keyword{internal}
