% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_descv_plots}
\alias{write_descv_plots}
\title{Write descriptive plots}
\usage{
write_descv_plots(
  data_tb,
  ds_descvs_ls,
  descv_outp_dir_1L_chr,
  lbl_nms_chr = c("Household tasks", "Getting around", "Morbility", "Self care",
    "Enjoy close rels", "Family rels", "Community involvement", "Despair", "Worry",
    "Sad", "Agitated", "Energy level", "Control", "Coping", "Frequency of pain",
    "Degree of pain", "Pain interference", "Vision", "Hearing", "Communication"),
  maui_domains_pfcs_1L_chr = "vD",
  item_plots_params_ls = list(plot_rows_cols_pair_int = c(5L, 4L), heights_int = c(10L,
    1L), width_1L_dbl = 9),
  dim_plots_params_ls = list(plot_rows_cols_pair_int = c(3L, 2L), heights_int = c(10L,
    1L), width_1L_dbl = 8),
  utl_by_rnd_plots_params_ls = list(width_1L_dbl = 6, height_1L_dbl = 4),
  combined_plot_params_ls = list(nrow_1L_int = 2L, rel_heights_dbl = c(4, 10),
    scale_dbl = c(0.9, 0.9), base_height_dbl = 10)
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{ds_descvs_ls}{Dataset descriptives (a list)}

\item{descv_outp_dir_1L_chr}{Descriptive output directory (a character vector of length one)}

\item{lbl_nms_chr}{Label names (a character vector), Default: c("Household tasks", "Getting around", "Morbility", "Self care",
"Enjoy close rels", "Family rels", "Community involvement",
"Despair", "Worry", "Sad", "Agitated", "Energy level", "Control",
"Coping", "Frequency of pain", "Degree of pain", "Pain interference",
"Vision", "Hearing", "Communication")}

\item{maui_domains_pfcs_1L_chr}{Maui domains pfcs (a character vector of length one), Default: 'vD'}

\item{item_plots_params_ls}{Item plots params (a list), Default: list(plot_rows_cols_pair_int = c(5L, 4L), heights_int = c(10L,
1L), width_1L_dbl = 9)}

\item{dim_plots_params_ls}{Dimension plots params (a list), Default: list(plot_rows_cols_pair_int = c(3L, 2L), heights_int = c(10L,
1L), width_1L_dbl = 8)}

\item{utl_by_rnd_plots_params_ls}{Utility by rnd plots params (a list), Default: list(width_1L_dbl = 6, height_1L_dbl = 4)}

\item{combined_plot_params_ls}{Combined plot params (a list), Default: list(nrow_1L_int = 2L, rel_heights_dbl = c(4, 10), scale_dbl = c(0.9,
0.9), base_height_dbl = 10)}
}
\value{
Descriptive plots paths (a list)
}
\description{
write_descv_plots() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write descriptive plots. The function returns Descriptive plots paths (a list).
}
\keyword{internal}
