% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_youthvars.R
\docType{package}
\name{youthvars-package}
\alias{youthvars}
\alias{youthvars-package}
\title{youthvars: Youth Mental Health Variables Description Toolkit}
\description{
Tools to describe, summarise and score types of data commonly
present in youth mental health collections.  The main motivation for
this package is to facilitate automated data integrity checks, ensure
that methods are applied to the appropriate data structures and
streamline reporting of descriptive statistics.  This development
version of the youthvars package has been made available as part of
the process of testing and documenting the package. The documentation
for this package has been automatically generated by the ready4fun
package and is therefore quite rudimentary. Human authored
documentation will follow in 2021.  If you have any questions, please
contact the authors (matthew.hamilton@orygen.org.au).
}
\details{
To learn more about youthvars, start with the vignettes:
\code{browseVignettes(package = "youthvars")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/youthvars/}
  \item \url{https://github.com/ready4-dev/youthvars}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Caroline Gao \email{caroline.gao@orygen.org.au} (\href{https://orcid.org/0000-0002-0987-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Headspace [funder]
  \item National Health and Medical Research Council [funder]
}

}
