% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_aqol6d_adol.R
\name{youthvars_aqol6d_adol}
\alias{youthvars_aqol6d_adol}
\alias{make_pt_youthvars_aqol6d_adol}
\alias{is_youthvars_aqol6d_adol}
\title{youthvars ready4 sub-module (S3 class) for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent)}
\usage{
youthvars_aqol6d_adol(x = make_pt_youthvars_aqol6d_adol())

make_pt_youthvars_aqol6d_adol()

is_youthvars_aqol6d_adol(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A validated instance of the youthvars ready4 sub-module (S3 class) for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent)

A prototype for youthvars ready4 sub-module (S3 class) for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent)

A logical value, TRUE if a valid instance of the youthvars ready4 sub-module (S3 class) for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent)
}
\description{
Create a new valid instance of the youthvars ready4 sub-module (S3 class) for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent)
}
\details{
youthvars ready4 sub-module (S3 class) for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent)
}
