% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_bads.R
\name{youthvars_bads}
\alias{youthvars_bads}
\title{First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores}
\usage{
youthvars_bads(x = make_pt_youthvars_bads())
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores, Default: make_pt_youthvars_bads()}
}
\value{
A validated instance of the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\description{
Create a new valid instance of the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\details{
First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
