% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_gad7.R
\name{youthvars_gad7}
\alias{youthvars_gad7}
\alias{make_pt_youthvars_gad7}
\alias{is_youthvars_gad7}
\title{youthvars S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores}
\usage{
youthvars_gad7(x = make_pt_youthvars_gad7())

make_pt_youthvars_gad7()

is_youthvars_gad7(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A validated instance of the youthvars S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores

A prototype for youthvars S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores

A logical value, TRUE if a valid instance of the youthvars S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\description{
Create a new valid instance of the youthvars S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\details{
youthvars S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
