% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_k6.R
\name{youthvars_k6}
\alias{youthvars_k6}
\title{youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores}
\usage{
youthvars_k6(x = make_pt_youthvars_k6())
}
\arguments{
\item{x}{A prototype for the youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores, Default: make_pt_youthvars_k6()}
}
\value{
A validated instance of the youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\description{
Create a new valid instance of the youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\details{
youthvars S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
