% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_oasis.R
\name{youthvars_oasis}
\alias{youthvars_oasis}
\title{youthvars S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores}
\usage{
youthvars_oasis(x = make_pt_youthvars_oasis())
}
\arguments{
\item{x}{A prototype for the youthvars S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores, Default: make_pt_youthvars_oasis()}
}
\value{
A validated instance of the youthvars S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\description{
Create a new valid instance of the youthvars S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\details{
youthvars S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
