% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_youthvars_scared.R
\name{youthvars_scared}
\alias{youthvars_scared}
\title{youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores}
\usage{
youthvars_scared(x = make_pt_youthvars_scared())
}
\arguments{
\item{x}{A prototype for the youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores, Default: make_pt_youthvars_scared()}
}
\value{
A validated instance of the youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\description{
Create a new valid instance of the youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\details{
youthvars S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
